Fix enum literal value and type reference md string

+++ ECMA2Yaml/ECMAHelper/Models/SDP/Shared/EnumField.cs
@@ -20,7 +20,7 @@ namespace ECMA2Yaml.Models.SDP
 
         [JsonProperty("literalValue")]
         [YamlMember(Alias = "literalValue")]
-        public int LiteralValue { get; set; }
+        public Int64 LiteralValue { get; set; }
 
         [JsonProperty("name")]
         [YamlMember(Alias = "name")]
